Add a threading scheduler
Create a scheduler that uses pthread to start and synchronize process
steps that run on threads. Creating the process creates a thread, writing
to its input will make it generate output. Blocks when either input is
empty or output is full. Allows interrupting the thread to allow it to
be terminated.