Add SingleElementQueue and add tests and code for arbitrary classes
SingleElementQueue is a queue that only has space for one element.
This makes it cheap for Cooperative processes to use.

Refactor test code to use typeof() to reduce code complexity.
Add tests for memory managrment of the queues and for demonstrating
what a class must support in order to become storable in a queue.