solver: Add a weighted optimizer (in test code)
Given a set of functions and bitmasks on which partitions they can be
implementeds, and a score for each function, calculate the distribution
that yields the optimal score.

Reduce the search space by calculating the maximum score that can be
attained for a certain node. If the current score is equal to that,
stop searching for anything better.