Add driver test for interrupt handling
Set up a thread that blocks on reading from a fifo queue.
Then write to the queue to unblock the thread. This verifies
that the IRQ hander is working correctly.

Write a fifo until it accepts no more data. Start a thread
and have it block on a "write" call. Then unblock the thread
by reading from the fifo.

Loop through all fifo's in the hardware_driver_poll test.