Split interrupt method into separate parts
When a process wants to stop, it interrupts its queues. However, this
will also affect any other process attached to those queues, because
that will also receive an InterruptedException when this happens,
and likely it will terminate. This is a problem when you want to
modify the chain, because the whole chain will stop processing.

To fix this issue, a process can ask the queue to interrupt only a
specific part, either "read" or "write". The other side remains
unaffected.