Split up "reset" into "clear" and "resume" methods.
Rename "reset" to "clear", its implementation has changed because
it no longer implies a "resume" on both ends of the queue.

Introduce "resume" methods that are balanced with the "interrupt"
methods introduced by the previous commit. This allows to stop/start
each end of a queue separately (if the scheduler allows it).

Refactor test names and group them into related topics.