Add usersignal support
The term "usersignal" means the extra data bytes that can be used
for signalling in logic. Since CPU communication is limited to 32
bit words, the extra bits are considered "out of band" data.

When a change in usersignal reaches the CPU or DMA node, it stops
the transfer to userspace, returning a short read. This can be used
to signal a frame change or end of stream condition, depending on the
application domain.