Add DMA reset and zero-copy block transfer capability
DMA reset stops all ongoing transfers in a node. This may cause data loss,
but is the only way to "stop" a DMA node and change its configuration.

To use zero-copy transfers, one has to open the DMA device and request
blocks to be allocated via ioctl calls. These blocks can be mapped into
the process and then queued into the driver. This works very similar to
what IIO does (in fact, it's based on that).

Add a dma_from_logic stress test now that reset() has been implemented.