hardware.cpp: No longer require "seek capable" input for programming bitstreams
This allows bitstreams to be programmed from stdin even when still in "bit"
format.

Unlike local files and pipes, network pipes heve a tendency to return "short"
reads, resulting in misaligned writes. Handle this in the File class by
introducing a "read_all" that keeps attempting to read more data until error
or end-of-file.