hardware: Allow performing UserID check. Faster byte-swap routine.
Add callbacks to inform the caller about the tags in a bitstream. This
makes it possible to check on UserID for example. Also add a parser
to extract UserID and PARTIAL properties from the description field.

Improve the byteswap routine by word-aligning it.

Use "read_all" when programming in binary mode too.