hardware: Simplify HardwareProgrammer
The HardwareProgrammer class does not need to open a new instance of
a control object, nor does it need to remember them. So it basically
becomes a wrapper for the DMA transfers.

In the test, use the ICAP method for initial programming too, so the
test can be run on a PCIe system.